requirements.txt
designer==0.6.6
pillow==10.4.0
pygame==2.6.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dotenv==1.0.1
